Fix some remaining CI issues
